+++ apps/commerce-api/src/main/resources/sql/insert_data.sql
@@ -23,3 +23,5 @@ INSERT INTO product (brand_id, name, price, stock_quantity, like_count, created_
 
 
 
+
+
